+++ react/typings/vtex.my-account-commons.d.ts
@@ -1,3 +1,4 @@
 declare module 'vtex.my-account-commons' {
   export const ContentWrapper
+  export const utils
 }